requirements.txt
discord.py==2.2.3
google_api_python_client==2.86.0
httplib2==0.22.0
oauth2client==4.1.3
ossapi==3.4.4
osu_sr_calculator==1.1.0
Pillow==9.5.0
pymongo==4.3.3
python-dotenv==1.0.0
python_dateutil==2.8.2
requests==2.25.1
schedule==1.2.0
numpy==1.24.3
